backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

